// packages/admin/admin/src/messages.d.ts
export declare const messages: {
    edit: {
        id: string;
        defaultMessage: string;
    };
    add: {
        id: string;
        defaultMessage: string;
    };
    saveUnsavedChanges: {
        id: string;
        defaultMessage: string;
        description: string;
    };
    save: {
        id: string;
        defaultMessage: string;
    };
    saveAndGoBack: {
        id: string;
        defaultMessage: string;
    };
    cancel: {
        id: string;
        defaultMessage: string;
    };
    delete: {
        id: string;
        defaultMessage: string;
    };
    ok: {
        id: string;
        defaultMessage: string;
    };
    unsavedChanges: {
        id: string;
        defaultMessage: string;
    };
    discard: {
        id: string;
        defaultMessage: string;
    };
    undo: {
        id: string;
        defaultMessage: string;
    };
    back: {
        id: string;
        defaultMessage: string;
    };
    reset: {
        id: string;
        defaultMessage: string;
    };
    apply: {
        id: string;
        defaultMessage: string;
    };
    copy: {
        id: string;
        defaultMessage: string;
    };
    paste: {
        id: string;
        defaultMessage: string;
    };
    left: {
        id: string;
        defaultMessage: string;
    };
    right: {
        id: string;
        defaultMessage: string;
    };
    link: {
        id: string;
        defaultMessage: string;
    };
    text: {
        id: string;
        defaultMessage: string;
    };
    file: {
        id: string;
        defaultMessage: string;
    };
    image: {
        id: string;
        defaultMessage: string;
    };
    page: {
        id: string;
        defaultMessage: string;
    };
    yes: {
        id: string;
        defaultMessage: string;
    };
    no: {
        id: string;
        defaultMessage: string;
    };
    globalContentScope: {
        id: string;
        defaultMessage: string;
    };
    invalidData: {
        id: string;
        defaultMessage: string;
    };
    saveConflict: {
        id: string;
        defaultMessage: string;
    };
    retry: {
        id: string;
        defaultMessage: string;
    };
    error: {
        id: string;
        defaultMessage: string;
    };
    networkError: {
        id: string;
        defaultMessage: string;
    };
    unknownError: {
        id: string;
        defaultMessage: string;
    };
    unknown: {
        id: string;
        defaultMessage: string;
    };
    warning: {
        id: string;
        defaultMessage: string;
    };
    content: {
        id: string;
        defaultMessage: string;
    };
    close: {
        id: string;
        defaultMessage: string;
    };
    url: {
        id: string;
        defaultMessage: string;
    };
    filter: {
        id: string;
        defaultMessage: string;
    };
    columns: {
        id: string;
        defaultMessage: string;
    };
    copyUrl: {
        id: string;
        defaultMessage: string;
    };
    deleteItem: {
        id: string;
        defaultMessage: string;
    };
    empty: {
        id: string;
        defaultMessage: string;
    };
    downloadAsExcel: {
        id: string;
        defaultMessage: string;
    };
    impersonate: {
        id: string;
        defaultMessage: string;
    };
    stopImpersonation: {
        id: string;
        defaultMessage: string;
    };
};
//# sourceMappingURL=messages.d.ts.map